+++ src/input_parsers/OnnxParser.cpp
@@ -65,8 +65,8 @@ OnnxParser::OnnxParser( const String &path )
 void OnnxParser::generateQuery( InputQuery& query )
 {
     Set<String> inputNames = readInputNames();
-    String outputName = readOutputName();
-    processGraph( inputNames, outputName, query );
+    Set<String> terminalNames = readOutputNames();
+    processGraph( inputNames, terminalNames, query );
 }
 
 /**
@@ -75,14 +75,14 @@ void OnnxParser::generateQuery( InputQuery& query )
  * @param query The query object to be populated
  * @param inputNames The set of input nodes. Note these don't have to be an actual input to the
  * network, they can be intermediate nodes.
- * @param outputName The output node. Note that again this doesn't have to be an actual output of
- * the network, it can be an intermediate node.
+ * @param terminalNames The set of terminal nodes. Note that these doesn't have to be outputs of
+ * the network, they can be intermediate nodes.
  */
-void OnnxParser::generatePartialQuery( InputQuery& query, Set<String>& inputNames, String& outputName )
+void OnnxParser::generatePartialQuery( InputQuery& query, Set<String>& inputNames, Set<String>& terminalNames )
 {
     validateUserInputNames( inputNames );
-    validateUserOutputNames( outputName );
-    processGraph( inputNames, outputName, query );
+    validateUserTerminalNames( terminalNames );
+    processGraph( inputNames, terminalNames, query );
 }
 
 /*************
@@ -126,6 +126,12 @@ void unimplementedConstantTypeError( onnx::TensorProto_DataType type )
 void unsupportedError( onnx::NodeProto &node )
 {
     String errorMessage = Stringf( "Onnx operation %s not currently supported by Marabou", node.op_type().c_str() );
+    throw MarabouError( MarabouError::ONNX_PARSER_ERROR, errorMessage.ascii() );
+}
+
+void unsupportedCastError ( onnx::TensorProto_DataType from, onnx::TensorProto_DataType to )
+{
+    String errorMessage = Stringf( "The ONNX parser does not currently support casting from '%s' to '%s'", TensorProto_DataType_Name(from).c_str(), TensorProto_DataType_Name(to).c_str());
     throw MarabouError( MarabouError::ONNX_PARSER_ERROR, errorMessage.ascii() );
 }
 
@@ -303,6 +309,16 @@ int getIntAttribute( onnx::NodeProto& node, String name, int defaultValue )
     if ( attr == nullptr )
     {
         return defaultValue;
+    }
+    return attr->i();
+}
+
+int getRequiredIntAttribute( onnx::NodeProto& node, String name )
+{
+    const onnx::AttributeProto* attr = findAttribute( node, name, onnx::AttributeProto_AttributeType_INT );
+    if ( attr == nullptr )
+    {
+        missingAttributeError( node, name );
     }
     return attr->i();
 }
@@ -494,19 +510,22 @@ void OnnxParser::validateUserInputNames( Set<String>& inputNames )
     }
 }
 
-void OnnxParser::validateUserOutputNames( String &outputName )
+void OnnxParser::validateUserTerminalNames( Set<String>& terminalNames )
 {
-    for ( auto node : _network.node() )
+    for ( String terminalName : terminalNames )
     {
-        for ( String outputNodeName : node.output() )
+        for ( auto node : _network.node() )
         {
-            if ( outputName == outputNodeName )
-                return;
+            for ( String outputNodeName : node.output() )
+            {
+                if ( terminalName == outputNodeName )
+                    return;
+            }
         }
-    }
 
-    String errorMessage = Stringf( "Output %s not found in graph!", outputName.ascii() );
-    throw MarabouError(  MarabouError::ONNX_PARSER_ERROR, errorMessage.ascii() );
+        String errorMessage = Stringf( "Output %s not found in graph!", terminalName.ascii() );
+        throw MarabouError(  MarabouError::ONNX_PARSER_ERROR, errorMessage.ascii() );
+    }
 }
 
 Set<String> OnnxParser::readInputNames()
@@ -530,22 +549,15 @@ Set<String> OnnxParser::readInputNames()
     return Set<String>::difference( inputNames, initializerNames );
 }
 
-String OnnxParser::readOutputName()
+Set<String> OnnxParser::readOutputNames()
 {
-    if ( _network.output().size() > 1 )
+    Set<String> outputNames;
+    for ( auto &outputNode : _network.output() )
     {
-        String message = "Your model has multiple outputs defined\n";
-        message += "Please specify the name of the output you want to consider using the 'outputName' argument\n";
-        message += "Possible options:";
-        for ( auto output : _network.output() )
-        {
-            message += " " + output.name();
-        }
-        throw MarabouError( MarabouError::ONNX_PARSER_ERROR, message.ascii() );
+        ONNX_LOG( Stringf( "Found output '%s'", outputNode.name().c_str() ).ascii() );
+        outputNames.insert( outputNode.name() );
     }
-    auto &outputNode = _network.output()[0];
-    ONNX_LOG( Stringf( "Found input: %s", outputNode.name().c_str() ).ascii() );
-    return outputNode.name();
+    return outputNames;
 }
 
 void OnnxParser::initializeShapeAndConstantMaps()
@@ -602,9 +614,12 @@ void OnnxParser::validateAllInputsAndOutputsFound()
     }
 
     // Mark the output variables
-    for ( Variable var : _varMap[_outputName] )
+    for ( String terminalName : _terminalNames )
     {
-        _outputVars.append( var );
+        for ( Variable var : _varMap[terminalName] )
+        {
+            _outputVars.append( var );
+        }
     }
 }
 
@@ -613,17 +628,20 @@ void OnnxParser::validateAllInputsAndOutputsFound()
  * Unlike the Python implementation, at the moment assumes there is only a single output.
  *
  * @param inputNames The names of the input nodes to start at.
- * @param outputName The names of the output node to end at.
+ * @param terminalNames The names of the output node to end at.
  * @param query The query in which to store the generated constraints.
  */
-void OnnxParser::processGraph( Set<String> &inputNames, String &outputName, InputQuery& query )
+void OnnxParser::processGraph( Set<String> &inputNames, Set<String> &terminalNames, InputQuery& query )
 {
     _inputNames = inputNames;
-    _outputName = outputName;
+    _terminalNames = terminalNames;
     _numberOfFoundInputs = 0;
 
     initializeShapeAndConstantMaps();
-    processNode( outputName, true );
+    for ( String terminalName : terminalNames )
+    {
+        processNode( terminalName, true );
+    }
     validateAllInputsAndOutputsFound();
     getMarabouQuery( query );
 }
@@ -848,9 +866,67 @@ void OnnxParser::identity( onnx::NodeProto& node )
  */
 void OnnxParser::cast( onnx::NodeProto& node )
 {
-    // See https://github.com/NeuralNetworkVerification/Marabou/blob/76b8eaf23518ca468c2cf05b742e3b4c858a64c3/maraboupy/MarabouNetworkONNX.py#L294
-    // for reference implementation
-    unimplementedOperationError( node );
+    String outputNodeName = node.output()[0];
+    String inputNodeName = node.input()[0];
+    _shapeMap[outputNodeName] = _shapeMap[inputNodeName];
+
+    // Try to find type to cast to. If not found, raise error.
+    onnx::TensorProto_DataType to = static_cast<onnx::TensorProto_DataType>( getRequiredIntAttribute( node, "to" ) );
+
+    if ( _varMap.exists(inputNodeName) )
+    {
+        // We shouldn't be casting variables to different types, since Marabou assumes variables
+        // have double precision
+        String errorMessage = Stringf( "The node '%s' casts non-constant values which is not supported by Marabou", inputNodeName.ascii() ) ;
+        throw MarabouError( MarabouError::ONNX_PARSER_ERROR, errorMessage.ascii() ) ;
+    }
+
+    if ( _constantIntTensors.exists( inputNodeName ) )
+    {
+        Vector<int> tensor = _constantIntTensors[inputNodeName];
+        if ( to == onnx::TensorProto_DataType_INT64 )
+        {
+            _constantIntTensors.insert( outputNodeName, tensor );
+        }
+        else if ( to == onnx::TensorProto_DataType_FLOAT )
+        {
+            Vector<double> castTensor = Vector<double>( tensor.size() );
+            for ( PackedTensorIndices i = 0; i < tensor.size(); i++ )
+            {
+                castTensor[i] = static_cast<double>( tensor[i] );
+            }
+            _constantFloatTensors.insert( outputNodeName, castTensor );
+        }
+        else
+        {
+            unsupportedCastError( onnx::TensorProto_DataType_INT64, to );
+        }
+    }
+    else if ( _constantFloatTensors.exists( inputNodeName ) )
+    {
+        Vector<double> tensor = _constantFloatTensors[inputNodeName];
+        if ( to == onnx::TensorProto_DataType_INT64 )
+        {
+            Vector<int> castTensor = Vector<int>( tensor.size() );
+            for ( PackedTensorIndices i = 0; i < tensor.size(); i++ )
+            {
+                castTensor[i] = static_cast<int>( tensor[i] );
+            }
+            _constantIntTensors.insert( outputNodeName, castTensor );
+        }
+        else if ( to == onnx::TensorProto_DataType_FLOAT )
+        {
+            _constantFloatTensors.insert( outputNodeName, tensor );
+        }
+        else
+        {
+            unsupportedCastError( onnx::TensorProto_DataType_FLOAT, to );
+        }
+    }
+    else
+    {
+        missingNodeError( inputNodeName );
+    }
 }
 
 /**